requirements.txt
fastapi==0.115.12
uvicorn==0.30.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
redis==4.5.1
websockets>=15.0.1
fastapi-limiter==0.1.5
pydantic==2.11.7
pydantic-settings==2.6.1
fastmcp
openai_harmony
typer==0.12.3
